+++ persistence/service_node.go
@@ -24,11 +24,11 @@ func (p PostgresContext) GetServiceNode(address []byte, height int64) (operator,
 	return
 }
 
-func (p PostgresContext) InsertServiceNode(address []byte, publicKey []byte, output []byte, _ bool, _ int, serviceURL string, stakedAmount string, chains []string, pausedHeight int64, unstakingHeight int64) error {
+func (p PostgresContext) InsertServiceNode(address []byte, publicKey []byte, output []byte, _ bool, _ int32, serviceURL string, stakedTokens string, chains []string, pausedHeight int64, unstakingHeight int64) error {
 	return p.InsertActor(types.ServiceNodeActor, types.BaseActor{
 		Address:            hex.EncodeToString(address),
 		PublicKey:          hex.EncodeToString(publicKey),
-		StakedTokens:       stakedAmount,
+		StakedTokens:       stakedTokens,
 		ActorSpecificParam: serviceURL,
 		OutputAddress:      hex.EncodeToString(output),
 		PausedHeight:       pausedHeight,
@@ -58,15 +58,15 @@ func (p PostgresContext) GetServiceNodeCount(chain string, height int64) (int, e
 	panic("GetServiceNodeCount not implemented")
 }
 
-func (p PostgresContext) GetServiceNodesReadyToUnstake(height int64, _ int) ([]modules.IUnstakingActor, error) {
+func (p PostgresContext) GetServiceNodesReadyToUnstake(height int64, status int32) ([]modules.IUnstakingActor, error) {
 	return p.GetActorsReadyToUnstake(types.ServiceNodeActor, height)
 }
 
-func (p PostgresContext) GetServiceNodeStatus(address []byte, height int64) (int, error) {
+func (p PostgresContext) GetServiceNodeStatus(address []byte, height int64) (int32, error) {
 	return p.GetActorStatus(types.ServiceNodeActor, address, height)
 }
 
-func (p PostgresContext) SetServiceNodeUnstakingHeightAndStatus(address []byte, unstakingHeight int64, _ int) error {
+func (p PostgresContext) SetServiceNodeUnstakingHeightAndStatus(address []byte, unstakingHeight int64, status int32) error {
 	return p.SetActorUnstakingHeightAndStatus(types.ServiceNodeActor, address, unstakingHeight)
 }
 
@@ -74,7 +74,7 @@ func (p PostgresContext) GetServiceNodePauseHeightIfExists(address []byte, heigh
 	return p.GetActorPauseHeightIfExists(types.ServiceNodeActor, address, height)
 }
 
-func (p PostgresContext) SetServiceNodeStatusAndUnstakingHeightIfPausedBefore(pausedBeforeHeight, unstakingHeight int64, _ int) error {
+func (p PostgresContext) SetServiceNodeStatusAndUnstakingHeightIfPausedBefore(pausedBeforeHeight, unstakingHeight int64, status int32) error {
 	return p.SetActorStatusAndUnstakingHeightIfPausedBefore(types.ServiceNodeActor, pausedBeforeHeight, unstakingHeight)
 }