+++ p2p/CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.5] - 2022-10-12
+
+### [#235](https://github.com/pokt-network/pocket/pull/235) Config and genesis handling
+
+- Updated to use `RuntimeMgr`
+- Updated tests and mocks
+- Removed some cross-module dependencies
+
+## [0.0.0.4] - 2022-10-06
+
+- Don't ignore the exit code of `m.Run()` in the unit tests
+
 ## [0.0.0.3] - 2022-09-15
 
 **[TECHDEBT] AddrBook management optimization and refactoring** [#246](github.com/pokt-network/pocket/issues/246)